+++ Assets/WorldGeneration/_Scripts/ScriptableObjects/AssetSettings.cs
@@ -1,0 +1,17 @@
+using System.Collections.Generic;
+using InGameTime;
+using UnityEngine;
+
+namespace WorldGeneration._Scripts.ScriptableObjects
+{
+    [CreateAssetMenu(menuName = "ScriptableObjects/Settings/Plant")]
+    public class AssetSettings : ScriptableObject
+    {
+        public GameObject assetPrefab;
+        public Transform parent;
+        public InGameDate lifespan;
+        public int maxNumber;
+
+        public List<GameObject> assets;
+    }
+}