+++ Assets/WorldGeneration/_Scripts/Spawning/AssetManager.cs
@@ -1,8 +1,8 @@
 using System;
+using System.Collections.Generic;
 using UnityEngine;
 using WorldGeneration._Scripts.Helper;
 using WorldGeneration._Scripts.ScriptableObjects;
-using WorldGeneration._Scripts.TerrainGeneration;
 using Random = UnityEngine.Random;
 
 namespace WorldGeneration._Scripts.Spawning
@@ -10,19 +10,98 @@ namespace WorldGeneration._Scripts.Spawning
     [Serializable]
     public class AssetManager
     {
+        public static AssetManager Instance;
+
         // private
-        private static AssetManager _instance;
-        private CarrotSpawner _carrotSpawner;
+        private List<AssetSettings> assets;
 
         public static AssetManager GetInstance()
         {
-            if (_instance == null)
+            if (Instance == null)
             {
-                _instance = new AssetManager();
-                return _instance;
+                Instance = new AssetManager();
+                return Instance;
             }
 
-            return _instance;
+            return Instance;
+        }
+
+        /// <summary>
+        /// </summary>
+        /// <param name="resolution"></param>
+        /// <param name="maxTerrainHeight"></param>
+        /// <param name="waterLevel"></param>
+        /// <param name="generalSettings"></param>
+        /// <param name="map"></param>
+        /// <param name="settings"></param>
+        /// <param name="parent"></param>
+        public void InitialSpawnAssets(Vector2Int resolution, float maxTerrainHeight, float waterLevel,
+            GeneralSettings generalSettings,
+            float[,] map, AssetSettings settings, Transform parent)
+        {
+            assets = new List<AssetSettings>();
+
+            for (var i = 0; i < settings.maxNumber; i++)
+            {
+                var posFound = false;
+                float xPos, zPos;
+                Vector3 pos;
+
+                while (!posFound)
+                {
+                    xPos = Random.Range(-generalSettings.squareSize * (resolution.x - 1) / 2,
+                        generalSettings.squareSize * (resolution.x - 1) / 2);
+                    zPos = Random.Range(-generalSettings.squareSize * (resolution.y - 1) / 2,
+                        generalSettings.squareSize * (resolution.y - 1) / 2);
+
+                    // zPos = Random.Range(0.0f, 1.0f) * (resolution.y - 1);
+                    // Debug.Log("xPos: " + xPos);
+                    // Debug.Log("zPos: "+ zPos);
+
+                    pos = GeneratorFunctions.GetSurfacePointFromWorldCoordinate(xPos, zPos, resolution,
+                        maxTerrainHeight,
+                        map, generalSettings);
+
+                    if (pos.y > maxTerrainHeight * waterLevel /*&& pos.y < maxTerrainHeight * 0.8f*/)
+                    {
+                        var asset = UnityEngine.Object.Instantiate(settings.assetPrefab, pos,
+                            Quaternion.Euler(new Vector3(0, Random.Range(0, 360), 0)));
+                        asset.transform.SetParent(parent);
+
+                        settings.assets.Add(asset);
+                        assets.Add(settings);
+
+                        posFound = true;
+                    }
+                }
+            }
+
+            // for (var i = 0; i < grassSettings.maxNumber; i++)
+            // {
+            //     // var posFound = false;
+            //     float xPos, zPos;
+            //     Vector3 pos;
+            //
+            //     xPos = Random.Range(-generalSettings.squareSize * (resolution.x - 1) / 2,
+            //         generalSettings.squareSize * (resolution.x - 1) / 2);
+            //     zPos = Random.Range(-generalSettings.squareSize * (resolution.y - 1) / 2,
+            //         generalSettings.squareSize * (resolution.y - 1) / 2);
+            //
+            //     // xPos = Random.Range(
+            //     //     -generalSettings.squareSize * resolution.x / 2 /* + generalSettings.squareSize / 2*/,
+            //     //     -generalSettings.squareSize * resolution.x / 2 + generalSettings.squareSize * resolution.x /*+
+            //     //     generalSettings.squareSize / 2*/);
+            //     // zPos = Random.Range(
+            //     //     -generalSettings.squareSize * resolution.y / 2 /* + generalSettings.squareSize / 2*/,
+            //     //     -generalSettings.squareSize * resolution.y / 2 + generalSettings.squareSize * resolution.y /*+
+            //     //     generalSettings.squareSize / 2*/);
+            //
+            //     pos = GeneratorFunctions.GetSurfacePointFromWorldCoordinate(xPos, zPos, resolution, maxTerrainHeight,
+            //         map, generalSettings);
+            //
+            //     var plant = _assetSpawner.Spawn(grassSettings.plantPrefab, pos);
+            //     plant.transform.SetParent(grassSettings.parent);
+            // }
         }
 
         /// <summary>
@@ -32,57 +111,60 @@ namespace WorldGeneration._Scripts.Spawning
         /// <param name="maxTerrainHeight"></param>
         /// <param name="waterLevel"></param>
         /// <param name="generalSettings"></param>
-        /// <param name="noiseSettings"></param>
         /// <param name="map"></param>
-        /// <param name="noiseWithClamp"></param>
-        /// <param name="demoCarrot"></param>
-        /// <param name="corners"></param>
-        /// <param name="maxNumberOfCarrots"></param>
+        /// <param name="settings"></param>
+        /// <param name="parent"></param>
+        /// <param name="count"></param>
         public void SpawnAssets(Vector2Int resolution, float maxTerrainHeight, float waterLevel,
             GeneralSettings generalSettings,
-            NoiseSettings noiseSettings, float[,] map, NoiseWithClamp noiseWithClamp, GameObject demoCarrot,
-            Vector2[] corners, int maxNumberOfCarrots)
+            float[,] map, AssetSettings settings, Transform parent, int count)
         {
-            _carrotSpawner = new CarrotSpawner();
-
-            for (var i = 0; i <= maxNumberOfCarrots; i++)
+            for (var i = 0; i < count; i++)
             {
                 var posFound = false;
                 float xPos, zPos;
                 Vector3 pos;
-
-                // xPos = 3.346446854546165441312716954234876589f;
-                // zPos = 6.023497893456012363497379756109274570f;
+                //     
+                //     xPos = Random.Range(-generalSettings.squareSize * (resolution.x - 1) / 2,
+                //         generalSettings.squareSize * (resolution.x - 1) / 2);
+                //     zPos = Random.Range(-generalSettings.squareSize * (resolution.y - 1) / 2,
+                //         generalSettings.squareSize * (resolution.y - 1) / 2);
                 //
-                // pos = GeneratorFunctions.GetSurfacePoint(xPos, zPos, resolution, maxTerrainHeight, generalSettings,
-                //     noiseSettings, noiseWithClamp);
+                //     pos = GeneratorFunctions.GetSurfacePointFromWorldCoordinate(xPos, zPos, resolution,
+                //         maxTerrainHeight, map, generalSettings);
                 //
-                // Debug.Log(pos.ToString());
-
+                //     var carrot = UnityEngine.Object.Instantiate(settings.assetPrefab, pos, Quaternion.Euler(new Vector3(0, Random.Range(0, 360), 0)));
+                //     carrot.transform.SetParent(parent);
+                //
+                //     settings.assets.Add(carrot);
+                // }
 
                 while (!posFound)
                 {
-                    xPos = Random.Range(0.0f, 1.0f) * (resolution.x - 1);
-                    zPos = Random.Range(0.0f, 1.0f) * (resolution.y - 1);
+                    xPos = Random.Range(-generalSettings.squareSize * (resolution.x - 1) / 2,
+                        generalSettings.squareSize * (resolution.x - 1) / 2);
+                    zPos = Random.Range(-generalSettings.squareSize * (resolution.y - 1) / 2,
+                        generalSettings.squareSize * (resolution.y - 1) / 2);
+
+                    // zPos = Random.Range(0.0f, 1.0f) * (resolution.y - 1);
                     // Debug.Log("xPos: " + xPos);
                     // Debug.Log("zPos: "+ zPos);
-                    
-                    pos = GeneratorFunctions.GetSurfacePoint(xPos, zPos, resolution, maxTerrainHeight, map, generalSettings
-                        /*noiseSettings, noiseWithClamp*/);
-                    if (pos.y > maxTerrainHeight * waterLevel && pos.y < maxTerrainHeight * 0.8f)
+
+                    pos = GeneratorFunctions.GetSurfacePointFromWorldCoordinate(xPos, zPos, resolution,
+                        maxTerrainHeight,
+                        map, generalSettings);
+
+                    if (pos.y > maxTerrainHeight * waterLevel /*&& pos.y < maxTerrainHeight * 0.8f*/)
                     {
-                        _carrotSpawner.Spawn(demoCarrot, pos);
+                        var asset = UnityEngine.Object.Instantiate(settings.assetPrefab, pos,
+                            Quaternion.Euler(new Vector3(0, Random.Range(0, 360), 0)));
+                        asset.transform.SetParent(parent);
+
+                        settings.assets.Add(asset);
+
                         posFound = true;
                     }
                 }
-
-
-                // _carrotSpawner.Spawn(demoCarrot,
-                //     GeneratorFunctions.GetSurfacePoint(0.0f, 0.0f, resolution, maxTerrainHeight, generalSettings,
-                //         noiseSettings, noiseWithClamp));
-                // _carrotSpawner.Spawn(demoCarrot,
-                //     GeneratorFunctions.GetSurfacePoint(resolution.x - 1, resolution.y - 1, resolution, maxTerrainHeight,
-                //         generalSettings, noiseSettings, noiseWithClamp));
             }
         }
     }